Stout: Fixed two bugs in `mkdtemp` that block Windows tests.
Currently there are two bugs in the Windows implementation of
`os::mkdtemp`, which together cause all agent tests to fail. We describe
these bugs below and explain the steps this commit will take to address
them.

The first concerns a mismatch between the POSIX specification of
`mkdtemp` and the Windows implementation of `os::mkdtemp`. Specifically,
the POSIX specification indicates that, if any component of the path
prefix in the directory pattern passed to `mkdtemp` does not exist, we
are to set `errno` to `ENOENT` and return a null pointer. In the Stout
POSIX wrapper, `os::mkdtemp`, we will return an `ErrnoError`. In
contrast, the Windows implementation will erroneously recursively create
all components of the path prefix. This commit will address this by
explicitly setting the `recursive` flag of the call to `os::mkdir` to
`false`.

The second concerns a failure to parse Unix-like paths when creating the
temporary directory. Specifically, both the POSIX and Windows
implementations of `os::mkdtemp` have a default argument: `/tmp/XXXXXX`.
When we replace the 'X' characters with random alphanumeric characters
and pass the result to `os::mkdir`, and when the `recursive` flag is
errorneously set to `true`, we will fail to parse the path on Windows,
because `os::mkdir` can only pass Windows-style paths, with the '\'
character. This causes all Agent tests to fail, e.g., when we try to
create sandbox directories.

Technically, this second issue is actually resolved by setting the
`recursive` flag to `false` in the call to `os::mkdir`, but here we
observe that `/tmp` is not the "right place" to put temporary files on
Windows anyway, and so we take the time to clean it up here, by
replacing the default string literal path `/tmp/XXXXXX` with the
platform-agnostic `path::join(os::temp(), "XXXXXX)`.

Review: https://reviews.apache.org/r/54415/
