+++ 3rdparty/stout/include/stout/os/posix/mkdtemp.hpp
@@ -23,7 +23,10 @@
 #include <string>
 
 #include <stout/error.hpp>
+#include <stout/path.hpp>
 #include <stout/try.hpp>
+
+#include <stout/os/temp.hpp>
 
 
 namespace os {
@@ -32,7 +35,8 @@ namespace os {
 // template. The template may be any path with _6_ `Xs' appended to
 // it, for example /tmp/temp.XXXXXX. The trailing `Xs' are replaced
 // with a unique alphanumeric combination.
-inline Try<std::string> mkdtemp(const std::string& path = "/tmp/XXXXXX")
+inline Try<std::string> mkdtemp(
+    const std::string& path = path::join(os::temp(), "XXXXXX"))
 {
   char* temp = new char[path.size() + 1];
   ::memcpy(temp, path.c_str(), path.size() + 1);
