+++ 3rdparty/stout/include/stout/os/windows/temp.hpp
@@ -1,0 +1,40 @@
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//  http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#ifndef __STOUT_OS_WINDOWS_TEMP_HPP__
+#define __STOUT_OS_WINDOWS_TEMP_HPP__
+
+#include <string>
+
+#include <stout/windows.hpp>
+
+
+namespace os {
+
+inline std::string temp()
+{
+  // Get temp folder for current user.
+  char temp_folder[MAX_PATH + 2];
+  if (::GetTempPath(MAX_PATH + 2, temp_folder) == 0) {
+    // Failed, try current folder.
+    if (::GetCurrentDirectory(MAX_PATH + 2, temp_folder) == 0) {
+      // Failed, use relative path.
+      return ".";
+    }
+  }
+
+  return std::string(temp_folder);
+}
+
+} // namespace os {
+
+#endif // __STOUT_OS_WINDOWS_TEMP_HPP__
